pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.11.11"
asyncio = "^3.4.3"
openai = "^1.61.0"
pydantic-settings = "^2.7.1"
aioredis = "^2.0.1"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
redis = "^5.2.1"
loguru = "^0.7.3"
